+++ java-security/src/test/java/com/sap/cloud/security/adapter/xs/XSUserInfoAdapterIntegrationTest.java
@@ -400,9 +400,12 @@ public class XSUserInfoAdapterIntegrationTest {
 				eq(clientCredentials),
 				eq(correctEnduserInfo.getAppToken()),
 				any(),
-				anyMap())).thenThrow(OAuth2ServiceException
-						.createWithStatusCodeAndResponseBody("Unauthorized", org.apache.http.HttpStatus.SC_UNAUTHORIZED,
-								"Unauthorized"));
+				anyMap()))
+						.thenThrow(OAuth2ServiceException.builder("Unauthorized")
+								.withStatusCode(org.apache.http.HttpStatus.SC_UNAUTHORIZED)
+								.withUri(tokenEndpointUri)
+								.build());
+
 		// execute flow
 		correctEnduserInfo.requestToken(request);
 	}