+++ token-client/src/main/java/com/sap/cloud/security/xsuaa/tokenflows/Cacheable.java
@@ -1,5 +1,7 @@
 package com.sap.cloud.security.xsuaa.tokenflows;
 
+import com.sap.cloud.security.config.CacheConfiguration;
+
 import javax.annotation.Nonnull;
 
 /**
@@ -9,8 +11,7 @@ import javax.annotation.Nonnull;
 public interface Cacheable {
 
 	/**
-	 * Provides the cache configuration of the component. Must not be null. Use
-	 * {@link CacheConfiguration#CACHE_DISABLED} to disable caching.
+	 * Provides the cache configuration of the component. Must not be null.
 	 *
 	 * @return the cache configuration
 	 */